requirements.txt
requests
pandas
sqlalchemy 
flask
python-dotenv